[makeotf_lib] use fixed width integer types from <stdint.h> to solve issues with 64-bit arch

+++ FDK/Tools/Programs/makeotf/makeotf_lib/source/cffread/cffread.c
@@ -5,6 +5,8 @@
  * Compact Font Format (CFF) parser. (Standalone library.)
  */
 
+#include <stdint.h>
+
 #include <string.h>
 #include <stdlib.h>
 #include <limits.h>
@@ -34,7 +36,7 @@ static short exenc[] = {
 #define ABS(v)            ((v) < 0 ? -(v) : (v))
 
 /* 16.16 fixed point arithmetic */
-typedef long Fixed;
+typedef int32_t Fixed;
 #define fixedScale    65536.0
 #define FixedMax    ((Fixed)0x7FFFFFFF)
 #define FixedMin    ((Fixed)0x80000000)
